Move ProductListAdapter class to productList package
